purchase status update on click

+++ client/src/components/Purchases/TableList.vue
@@ -9,13 +9,41 @@
       <template slot="items" slot-scope="props">
           <td class="text-xs-center">{{ props.item.nfe }}</td>
           <td class="text-xs-center">{{ props.item.issueDate | formatedDate }}</td>
-          <td class="text-xs-left">{{ props.item.status }}</td>
+          <td class="text-xs-center" style="width: 120px">
+            <v-menu offset-y nudge-left="25%" full-width class="n-menu">
+              <v-chip
+                slot="activator"
+                class="n-chip"
+                :color="statusColor(props.item.status)"
+                small
+                block
+                outline
+              >{{ statusText(props.item.status) }}
+              </v-chip>
+              <v-list>
+                <v-list-tile
+                  v-for="(status, i) in purchaseStatus"
+                  :key="i"
+                  @click="editPurchaseStatus(props.item, status.value)"
+                >
+                  <v-chip
+                    class="n-chip"
+                    :color="statusColor(status.value)"
+                    small
+                    outline
+                  >{{ status.name }}</v-chip>
+                </v-list-tile>
+              </v-list>
+            </v-menu>
+          </td>
           <td class="text-xs-center">{{ props.item.total | currencyBRL }}</td>
-          <td>
+          <td style="width: 120px">
             <v-layout row>
-              <v-icon @click="editItem(props.item)" v-text="`edit`"></v-icon>
+              <v-icon @click="editItem(props.item)" v-text="`visibility`" small></v-icon>
               <v-spacer></v-spacer>
-              <v-icon  @click="deleteItem(props.item)" v-text="`delete`"></v-icon>
+              <v-icon @click="editItem(props.item)" v-text="`edit`" small></v-icon>
+              <v-spacer></v-spacer>
+              <v-icon  @click="deleteItem(props.item)" v-text="`delete`" small v-show="props.item.status !== 'confirmed'"></v-icon>
             </v-layout>
           </td>
       </template>
@@ -36,13 +64,53 @@ export default {
     headers: [
       { text: 'NFE', sortable: false, value: 'nfe' },
       { text: 'Data de emissão', align: 'center', sortable: false, value: 'issueDate' },
-      { text: 'status', align: 'left', sortable: false, value: 'status' },
+      { text: 'Status', align: 'center', sortable: false, value: 'status' },
       { text: 'Valor', align: 'center', sortable: false, value: 'value' },
       { text: 'Ações', align: 'center', sortable: false, value: 'actions' }
+    ],
+    purchaseStatus: [
+      { 'name': 'cancelado', value: 'canceled' },
+      { 'name': 'pendente', value: 'pending' },
+      { 'name': 'confirmado', value: 'confirmed' }
     ]
   }),
 
   methods: {
+    statusColor(status) {
+      switch (status) {
+        case 'confirmed':
+          return 'success'
+          break
+        case 'pending':
+          return 'warning'
+          break
+        case 'canceled':
+          return 'error'
+          break
+      }
+    },
+    statusText(status) {
+      switch (status) {
+        case 'confirmed':
+          return 'confirmado'
+          break
+        case 'pending':
+          return 'pendente'
+          break
+        case 'canceled':
+          return 'cancelado'
+          break
+      }
+    },
+    visualizePurchase (item) {
+      this.$emit('visualizePurchase', item)
+    },
+    editPurchaseStatus (item, status) {
+      // console.log(item, status)
+      const { _id } = item
+      const payload = { _id, status }
+      this.$emit('editPurchaseStatus', payload)
+    },
     editItem (item) {
       this.$emit('editItem', item)
     },
@@ -54,5 +122,14 @@ export default {
 </script>
 
 <style>
+.n-chip {
+  cursor: pointer;
+  width: 100%!important;
+  display: flex;
+  justify-content: center;
+}
 
+.n-chip .v-chip__content {
+  cursor: pointer;
+}
 </style>
